refactor: update installation scripts and fix successMsg typos

+++ scripts/05-script.sh
@@ -11,13 +11,13 @@ if [ $? -ne 0 ]; then
     clear
     exit 1
 else
-    echo "$sucessMsg"
+    echo "$successMsg"
 fi
 }
 #These variables need to be set directly after a process ends to capture the $? value and output a message, cmdFail runs function.
 #exitStat=$?
 #errMsg="ERROR MESSAGE"
-#sucessMsg="SUCCESS MESSAGE"
+#successMsg="SUCCESS MESSAGE"
 #cmdFail
 echo "Setting up Folder Variables"
 echo
@@ -61,18 +61,21 @@ sleep 0.5
 intelGPU=$(lspci | grep -i vga | grep -i "Intel")
 exitStat=$?
 errMsg="Error on Intel GPU check"
-sucessMsg="Intel GPU check passed"
+successMsg="Intel GPU check passed"
 cmdFail
 amdGPU=$(lspci | grep -i vga | grep -i "AMD")
 exitStat=$?
 errMsg="Error on AMD GPU check"
-sucessMsg="AMD GPU check passed"
+successMsg="AMD GPU check passed"
 cmdFail
 nvidiaGPU=$(lspci | grep -i vga | grep -i "NVIDIA")
 exitStat=$?
 errMsg="Error on NVIDIA GPU check"
-sucessMsg="NVIDIA GPU check passed"
+successMsg="NVIDIA GPU check passed"
 cmdFail
+installIntel=false
+installAMD=false
+installNVIDIA=false
 if [ -n "$intelGPU" ]; then
     echo
     echo "Intel GPU detected"
@@ -98,47 +101,47 @@ else
     echo
     echo "No GPU detected"
 fi
-depIntel=$(cat /opt/kevrevrun/cfg/deps//intel.apt)
- if [ $installIntel == true ]; then
+if [ $installIntel == true ]; then
+    depIntel=$(cat /opt/kevrevrun/cfg/deps/intelgpu.apt)
     echo
     echo "Installing Intel GPU drivers"
     sleep 1
     sudo DEBIAN_FRONTEND=noninteractive apt install $depIntel -y 2>&1
     exitStat=$?
     errMsg="Failed to install Intel GPU drivers"
-    sucessMsg="Intel GPU drivers installed successfully"
+    successMsg="Intel GPU drivers installed successfully"
     cmdFail
+    echo "Intel driver installation completed successfully"
+    read -p "Press [ENTER] key to continue..."
 fi
-echo "Intel driver installation completed successfully"
-read -p "Press [ENTER] key to continue..."
-depAMD=$(cat /opt/kevrevrun/cfg/deps/amdgpu.apt)
 if [ $installAMD == true ]; then
+    depAMD=$(cat /opt/kevrevrun/cfg/deps/amdgpu.apt)
     echo
     echo "Installing AMD GPU drivers"
     sleep 1
     sudo DEBIAN_FRONTEND=noninteractive apt install $depAMD -y 2>&1
     exitStat=$?
     errMsg="Failed to install AMD GPU drivers"
-    sucessMsg="AMD GPU drivers installed successfully"
+    successMsg="AMD GPU drivers installed successfully"
     cmdFail
     echo "AMD GPU driver installation completed successfully"
     read -p "Press [ENTER] key to continue..."
 fi
-depNVIDIA=$(cat /opt/kevrevrun/cfg/deps/nvidia.apt)
-urlNVIDIA=$(cat /opt/kevrevrun/cfg/nvidia.url)
 if [ $installNVIDIA == true ]; then
+    depNVIDIA=$(cat /opt/kevrevrun/cfg/deps/nvidia.apt)
+    urlNVIDIA=$(cat /opt/kevrevrun/cfg/nvidia.url)
     echo
     echo "Adding Nvidia Driver Repository"
     sleep 1
     wget -nv -O $tmpDir/cuda.deb $urlNVIDIA
     exitStat=$?
     errMsg="Failed to download Nvidia Driver Repository"
-    sucessMsg="Nvidia Driver Repository downloaded successfully"
+    successMsg="Nvidia Driver Repository downloaded successfully"
     cmdFail
     sudo dpkg -i $tmpDir/cuda.deb
     exitStat=$?
     errMsg="Failed to install Nvidia Driver Repository"
-    sucessMsg="Nvidia Driver Repository installed successfully"
+    successMsg="Nvidia Driver Repository installed successfully"
     cmdFail
     echo
     echo "Updating APT package cache"
@@ -147,7 +150,7 @@ if [ $installNVIDIA == true ]; then
     sudo DEBIAN_FRONTEND=noninteractive apt update
     exitStat=$?
     errMsg="Failed to update APT package cache"
-    sucessMsg="APT package cache updated successfully"
+    successMsg="APT package cache updated successfully"
     cmdFail
     echo
     echo "Cleaning up temporary files"
@@ -156,7 +159,7 @@ if [ $installNVIDIA == true ]; then
     rm -fv $tmpDir/cuda.deb
     exitStat=$?
     errMsg="Failed to remove temporary files"
-    sucessMsg="Temporary files removed successfully"
+    successMsg="Temporary files removed successfully"
     cmdFail
     echo
     echo "Installing Nvidia Dependancies"
@@ -165,7 +168,7 @@ if [ $installNVIDIA == true ]; then
     sudo DEBIAN_FRONTEND=noninteractive apt install $depNVIDIA -y
     exitStat=$?
     errMsg="Failed to install Nvidia Dependancies"
-    sucessMsg="Nvidia Dependancies installed successfully"
+    successMsg="Nvidia Dependancies installed successfully"
     cmdFail
     echo
     echo "Installing NVIDIA Driver Packages"
@@ -174,7 +177,7 @@ if [ $installNVIDIA == true ]; then
     sudo DEBIAN_FRONTEND=noninteractive apt install nvidia-open -y
     exitStat=$?
     errMsg="Failed to install NVIDIA Driver Packages"
-    sucessMsg="NVIDIA Driver installation completed successfully"
+    successMsg="NVIDIA Driver installation completed successfully"
     cmdFail
     # The package is signed if secure boot it the key may need to be added via mokutil. Will try install on Nvidia system to confirm if it is needed.
     echo
@@ -193,5 +196,5 @@ sleep 1
 echo 
 read -p "Press [Enter] key when ready to reboot..."
 clear
-reboot
+#sudo reboot
 exit 0
